Refactor code into more files &  clean up
